\name{GBM}
\docType{data}
\alias{GBM}
\title{list of 3 TCGA glioblastoma samples, parts thereof, outputs of SCANVISscan and SCANVISlinkvar functions with toy variants supplied for the variant-SJ mapping}
\description{list of parts of 3 matrices output by SCANVISscan}
\usage{GBM}
\format{list}
\keyword{datasets}
\keyword{SCANVISexamples}
