\name{gbm3}
\docType{data}
\alias{gbm3}
\title{part of a TCGA glioblastoma sample from STAR alignment SJ.tab file}
\description{matrix with chr,start,end,uniq.reads indicating genomic coordinates
for each splice junction and corresonding number of supporting split reads}
\usage{gbm3}
\format{matrix}
\keyword{datasets}
\keyword{SCANVISexamples}
