\name{gbm3.vcf}
\docType{data}
\alias{gbm3.vcf}
\title{a toy set of 6 variants that pair up with the gbm3 data example}
\description{matrix with chr,start,end,passedMUT indicating genomic coordinates
for each toy variant that has supposedly passed some threshold}
\usage{gbm3}
\format{matrix}
\keyword{datasets}
\keyword{SCANVISexamples}
