\name{gen19}
\docType{data}
\alias{gen19}
\title{parts of the annotation object created by the SCANVISannotation function
when used with the url ftp://ftp.ebi.ac.uk/pub/databases/gencode/Gencode_human/release_19/
which references the gencode v19 GTF file for human hg19}
\description{list with various genomic components for use in SCANVISfunctions}
\usage{gen19}
\format{list}
\keyword{datasets}
\keyword{SCANVISexamples}
\keyword{gencode}
\keyword{annotation}
\keyword{SCANVISannotation}
