\name{LUSC}
\docType{data}
\alias{LUSC}
\title{list of 3 TCGA lung squamous cell carcinoma samples, parts thereof, both outputs of SCANVISscan with the second sample
being variant-mapped via SCANVISlinkvar}
\description{list of parts of 3 matrices as output by SCANVISscan}
\usage{LUSC}
\format{list}
\keyword{datasets}
\keyword{SCANVISexamples}
