\name{SCANVIS-package}
\alias{SCANVIS-package}
\alias{SCANVIS}
\docType{package}
\title{
\packageTitle{SCANVIS}
}
\description{
\packageDescription{SCANVIS}
}
\details{

\packageDESCRIPTION{SCANVIS}
\packageIndices{SCANVIS}
SCANVIS is a set of tools for SCoring and ANnotating splice junctions using gencode annotation. It also has a VISualization component that allows users to quickly view one or more samples in sashimi style plots, showing splice junctions (SJs) and, optionally, a read coverage profile as well as mutations in one figure. These sashimi style plots are novel in that unannotated splice junctions are highlighted in various colours to delineate various junction types, with line styles indicating whether unannotated junctions are in frame or not.

}
%\license{SOFTWARE COPYRIGHT NOTICE AGREEMENT: This software and its documentation are copyright (2019) by the New York Genome Center. All rights are reserved. This software is supplied without any warranty or guaranteed support whatsoever. The New York Genome Center cannot be responsible for its use,misuse, or functionality.}

\author{
\packageAuthor{SCANVIS}

Maintainer: \packageMaintainer{SCANVIS}
}
%\references{Sashimi style plots inspired by Miso software (https://doi.org/10.1038/nmeth.1528)}

%~~ Optionally other standard keywords, one per line, from file KEYWORDS in ~~
%~~ the R documentation directory ~~
%\keyword{ package }
%\seealso{
%~~ Optional links to other man pages, e.g. ~~
%~~ \code{\link[<pkg>:<pkg>-package]{<pkg>}} ~~
%}
