\name{SCANVIS_examples}
\docType{data}
\alias{SCANVIS_data}
\title{Data for running SCANVIS examples}
\description{
  This data set includes the following:
   gencode19 (hg19) annotation details as derived by SCANVIS.annotation
   SJ details derived from STAR hg19 alignments to GBM/LUSC/LUAD samples from TCGA}
\usage{SCANVIS_examples}
\format{Contains the following data pieces: GBM, LUAD, LUSC, gbm3, gbm3.vcf}
\keyword{datasets,GBM,LUAD,LUSC,gbm3,gbm3.vcf}
