\name{GBM}
\docType{data}
\alias{GBM}
\title{list of 3 TCGA glioblastoma samples, parts thereof, outputs of SCANVIS.scan and SCANVIS.linkvar functions with toy variants supplied for the variant-SJ mapping}
\description{list of parts of 3 matrices output by SCANVIS.scan}
\usage{GBM}
\format{list}
\keyword{datasets}
\keyword{SCANVIS_examples}
