\name{LUSC}
\docType{data}
\alias{LUSC}
\title{list of 3 TCGA lung squamous cell carcinoma samples, parts thereof, both outputs of SCANVIS.scan with the second sample
being variant-mapped via SCANVIS.linkvar}
\description{list of parts of 3 matrices as output by SCANVIS.scan}
\usage{LUSC}
\format{list}
\keyword{datasets}
\keyword{SCANVIS_examples}
