\name{SCANVIS_examples}
\docType{data}
\alias{SCANVIS_examples}
\title{Data for running SCANVIS examples}
\description{This data is a list of parts of SJ.tab files as output by the STAR
alignment algorithm on a number of TCGA samples (GBM, LUSC and LUAD). It also 
includes part of the annotation file as derived by SCANVIS.annotation on hg19,
gencode v19 that was used to generate the included examples.} 
\usage{SCANVIS_examples}
\format{Contains the following data pieces: GBM, LUAD, LUSC, gbm3, gbm3.vcf,gen19}
\keyword{datasets}
