\name{SCANVIS_examples}
\docType{data}
\alias{SCANVIS_data}
\title{Data for running SCANVIS examples}
\description{
  This data set includes SJ details derived from STAR hg19 alignments to a number of TCGA samples}
\usage{SCANVIS_examples}
\format{Contains the following data pieces: GBM, LUAD, LUSC, gbm3, gbm3.vcf}
\keyword{datasets,GBM,LUAD,LUSC,gbm3,gbm3.vcf}
